Implemented some more versioning
git-svn-id: https://svn.kuali.org/repos/student/branches/ks-1.1-ver@16000 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ks-core/ks-core-api/src/main/java/org/kuali/student/core/versionmanagement/service/VersionManagementService.java
@@ -106,7 +106,7 @@ public interface VersionManagementService {
      * @throws PermissionDeniedException
      *             authorization failure
      */
-    public VersionDisplayInfo getVersionBySequenceNumber(@WebParam(name = "refObjectTypeURI") String refObjectTypeURI, @WebParam(name = "refObjectId") String refObjectId, @WebParam(name = "sequence") Integer sequence) throws DoesNotExistException, InvalidParameterException, MissingParameterException, OperationFailedException, PermissionDeniedException;
+    public VersionDisplayInfo getVersionBySequenceNumber(@WebParam(name = "refObjectTypeURI") String refObjectTypeURI, @WebParam(name = "refObjectId") String refObjectId, @WebParam(name = "sequence") Long sequence) throws DoesNotExistException, InvalidParameterException, MissingParameterException, OperationFailedException, PermissionDeniedException;
 
 
     /**
